requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Mahjong library
mahjong>=1.2.0

# Data processing
pyyaml>=6.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# XML parsing (already in standard library)
# lxml>=4.9.0  # Optional, for faster XML parsing

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
flake8>=6.0.0
black>=23.0.0
mypy>=1.4.0

# Logging
tensorboard>=2.13.0  # Optional, for TensorBoard logging

# Optional: For specific features
# requests>=2.31.0  # For downloading data
# pillow>=10.0.0  # For image processing
# opencv-python>=4.8.0  # For advanced visualizations

